rename: 첫 화면 컴포넌트 이름 수정
